pyproject.toml
[project]
name = "tg-bot-gpt"
version = "0.1.0"
description = "JavaRush first project"
authors = [
    {name = "littlebox-spb",email = "altukhov.alexandr@mail.ru"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai (>=1.82.0,<2.0.0)",
    "python-telegram-bot (>=22.1,<23.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "pytelegrambotapi (>=4.27.0,<5.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
